requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
black==22.8.0
cached-property==1.5.2
celery==5.1.2
certifi==2022.12.7
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
dataclasses==0.8
Django==3.2.19
django-celery-beat==2.2.1
django-crontab==0.7.1
django-elasticsearch-dsl==7.3
django-rest-framework==0.1.0
django-timezone-field==4.2.3
djangorestframework==3.14.0
elastic-transport==8.4.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
Flask==2.0.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.1.0
MarkupSafe==2.0.1
mypy-extensions==1.0.0
numpy==1.19.5
packaging==21.3
path==16.2.0
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.36
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyparsing==3.0.7
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.3.6
six==1.16.0
sqlparse==0.4.3
typed-ast==1.5.4
typing_extensions==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.0.3
zipp==3.6.0
